requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.8
django-filter==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
Markdown==2.6.11
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
urllib3==1.23
